test: reorganize BookmarkPosition specs
