+++ tests/unit/domain/value-objects/BookmarkPosition/test-utils.ts
@@ -1,0 +1,3 @@
+export const validSurahId = 1;
+export const validAyahNumber = 1;
+export const validTimestamp = new Date('2024-01-01T10:00:00Z');
